requirements.txt
Flask==2.3.3
python-chess==1.999
stockfish==3.28.0
requests==2.31.0
numpy==1.24.3
scipy==1.10.1
plotly==5.17.0
pandas==2.0.3
python-dateutil==2.8.2
Werkzeug==2.3.7 